requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
branca==0.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cvzone==1.6.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.8
et_xmlfile==2.0.0
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
flatbuffers==24.3.25
folium==0.19.0
fonttools==4.51.0
fsspec==2025.3.0
gast==0.6.0
geopandas==1.0.1
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.1
h5py==3.11.0
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.26.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kagglehub==0.3.4
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
missingno==0.5.2
ml-dtypes==0.4.1
mlxtend==0.23.3
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.14.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==1.0.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.1.2
pyproj==3.7.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.8.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==75.1.0
sgmllib3k==1.0.0
shapely==2.0.6
sip==6.8.6
six==1.16.0
sounddevice==0.5.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.6
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xyzservices==2024.9.0
